pyproject.toml
[project]
name = "jobmatch-assistant"
version = "1.0.0"
description = "AI-powered job application assistant using GPT-4o-mini, LangChain, Pinecone, and FastAPI for intelligent resume matching and cover letter generation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0",
    "alembic>=1.17.1",
    "psycopg[binary]>=3.2.12",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "bcrypt==4.2.1",
    "pydantic[email]>=2.12.3",
    "pydantic-settings>=2.0.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-openai>=1.0.1",
    "langchain-pinecone>=0.2.13",
    "openai>=2.6.1",
    "tiktoken>=0.12.0",
    "pdfplumber>=0.11.7",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "streamlit>=1.51.0",
    "slowapi>=0.1.9",
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
    "python-magic>=0.4.27; sys_platform != 'win32'",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
# Don't treat these directories as packages
norecursedirs = [".git", ".venv", "__pycache__", "*.egg-info", "dist", "build"]
